// src/app/Model/addToCart.ts
export interface AddToCart {
  pCustomerCode: string;
  pProjectCode: string;
  pOrderType: string;
  pOrderNo: number;
  pRefNo: number;
  pStructureElement: string;
  pProductType: string;
  pWBS1: string;
  pWBS2: string;
  pWBS3: string;
  pPONo: string;
  pScheduledProd: string;
  pPostID: number;
  UpdateBy: string;
  GreenSteel: boolean | null;
  AddressCode: string
}